+++ es/processors/loaders/FileReader.js
@@ -5,8 +5,6 @@ const debug = require('debug')('jason:load:file');
 
 const readFileAsync = promisify(fs.readFile);
 
-const REGEX_PAGINATION_EXP = /{(.+)}/;
-
 /**
  * Reads the content of a file.
  */
@@ -18,29 +16,14 @@ class FileReader {
    * @param {string} [config.encoding='utf8']
    */
   constructor({ config }) {
-    this._readConfig = {};
-    this._config = {};
-
-    Object.keys(config).forEach((key) => {
-      if (key[0] !== '_') {
-        this._readConfig[key] = config[key];
-      } else {
-        this._config[key.slice(1)] = config[key];
-      }
-    });
-
-    this._readConfig = {
+    this._config = {
       basePath: process.cwd(),
       stream: false,
       encoding: 'utf8',
-      ...this._readConfig,
+      ...config,
     };
 
-    this._config = { concurrency: 1, ...this._config };
-    this._config.concurrency = Number(this._config.concurrency); // just in case
-
     debug('FileReader instance created.');
-    debug('read config', this._readConfig);
     debug('config', this._config);
   }
 
@@ -50,42 +33,6 @@ class FileReader {
    */
   getConfig() {
     return this._config;
-  }
-
-  /**
-   * Builds all the links defined by the pagination config.
-   * @return {Array}
-   */
-  buildPaginationLinks() {
-    const configPath = this._readConfig.path || '';
-
-    const matches = configPath.match(REGEX_PAGINATION_EXP);
-    if (!matches) {
-      return [configPath];
-    }
-
-    let [start, end] = matches[1].split(',').map(s => Number(s));
-
-    if (end === undefined) {
-      end = start;
-    } else if (Number.isNaN(end) || end < start) {
-      debug('Warning: invalid end value for pagination ("%s")! Forcing to %s.', end, start);
-      end = start;
-    }
-
-    debug('Building pagination from %d -> %d.', start, end);
-
-    const links = [];
-
-    while (start <= end) {
-      const path = configPath.replace(REGEX_PAGINATION_EXP, start);
-      links.push(path);
-      start += 1;
-    }
-
-    debug(links);
-
-    return links;
   }
 
   /**
@@ -103,8 +50,8 @@ class FileReader {
    * @param {Object} [options] Optional read options.
    * @return {Promise}
    */
-  async run({ options }) {
-    const readConfig = { ...this._readConfig, ...options };
+  async run({ options } = {}) {
+    const readConfig = { ...this._config, ...options };
 
     const {
       basePath,